ZC QD Test: Combine Direct API and Readonly QD test in one program
The basic structure of QD tests remains the same and the same program
can be used for the newer QD tests that are yet to be added for ZC EM
API and ZC Bcast API.
