--- tests/charm++/zerocopy/zerocopy_with_qd/large_readonly/ro_qd_test.ci
@@ -1,20 +1,0 @@
-mainmodule ro_qd_test {
-
-  readonly int arr_size;
-  readonly int vec_size;
-
-  readonly int num_arr1[2000000];
-  readonly std::vector<int> num_vec1;
-
-  readonly CProxy_Main mProxy;
-
-  mainchare Main{
-    entry Main(CkArgMsg *msg);
-    entry [reductiontarget] void done();
-    entry void qdReached();
-  };
-
-  array [1D] arr1 {
-    entry arr1();
-  }
-};
