+++ tests/charm++/zerocopy/zerocopy_with_qd/zerocopy_with_qd.ci
@@ -1,10 +1,17 @@
 mainmodule zerocopy_with_qd {
 
+  readonly int arr_size;
+  readonly int vec_size;
+
+  readonly int num_arr1[2000000];
+  readonly std::vector<int> num_vec1;
+
   readonly int numElements;
   readonly CProxy_Main mProxy;
 
   mainchare Main{
     entry Main(CkArgMsg *m);
+    entry [reductiontarget] void done();
     entry void qdReached();
     entry void zcSrcCompleted(CkDataMsg *m);
     entry void zcDestCompleted(CkDataMsg *m);
@@ -12,6 +19,8 @@ mainmodule zerocopy_with_qd {
 
   array [1D] testArr{
     entry testArr();
+
+    entry void testDirectApi();
     entry void recvBufferInfo(CkNcpyBuffer src1, CkNcpyBuffer src2, CkNcpyBuffer src3);
   };
 };
